+++ R/Field.R
@@ -33,7 +33,7 @@
 #' field$name
 #'
 #' # Set the new data type
-#' field$datatype = pl$Categorical
+#' field$datatype = pl$Categorical()
 #' field$datatype
 #'
 #' # Set the new name