+++ R/extendr-wrappers.R
@@ -242,6 +242,8 @@ RPolarsDataType <- new.env(parent = emptyenv())
 
 RPolarsDataType$new <- function(s) .Call(wrap__RPolarsDataType__new, s)
 
+RPolarsDataType$new_categorical <- function(ordering) .Call(wrap__RPolarsDataType__new_categorical, ordering)
+
 RPolarsDataType$new_datetime <- function(tu, tz) .Call(wrap__RPolarsDataType__new_datetime, tu, tz)
 
 RPolarsDataType$new_duration <- function() .Call(wrap__RPolarsDataType__new_duration)