+++ tests/testthat/test-datatype.R
@@ -18,7 +18,7 @@ test_that("plStruct", {
   )
   expect_false(
     pl$Struct(bin = pl$Binary, pl$Field("bool", pl$Boolean)) ==
-      pl$Struct(list(bin = pl$Categorical, bool = pl$Boolean))
+      pl$Struct(list(bin = pl$Categorical(), bool = pl$Boolean))
   )
 
   # this would likely cause an error at query time though
@@ -69,4 +69,18 @@ test_that("String and Utf8 are identical", {
   string = pl$DataFrame(x = "a", schema = list(x = pl$String))$to_data_frame()
   utf8 = pl$DataFrame(x = "a", schema = list(x = pl$Utf8))$to_data_frame()
   expect_identical(string, utf8)
+})
+
+test_that("Categorical", {
+  expect_identical(
+    pl$Series(c("z", "z", "k", "a"))$cast(pl$Categorical())$sort()$to_r(),
+    factor(c("z", "z", "k", "a"))
+  )
+  expect_identical(
+    pl$Series(c("z", "z", "k", "a"))$cast(pl$Categorical("lexical"))$sort()$to_r(),
+    factor(c("a", "k", "z", "z"))
+  )
+  expect_error(
+    pl$Series(c("z", "z", "k", "a"))$cast(pl$Categorical("foobar"))
+  )
 })